feat: Função colocada em arquivo ts.
