+++ src/pages/despesas/Despesas.tsx
@@ -1,20 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import './Despesas.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
-
+import {calculateDaysLeft} from './calculateDaysLeft';
 export default function Despesas() {
   const [daysLeft, setDaysLeft] = useState(0);
-
-  // Função para calcular os dias restantes até a data de competência
-  const calculateDaysLeft = (endDate: string) => {
-    const today = new Date(); // Data atual
-    const [day, month, year] = endDate.split('/').map(Number); // Quebra a data no formato DD/MM/YYYY
-    const end = new Date(year, month - 1, day); // Cria o objeto Date com a data de competência
-
-    const differenceInTime = end.getTime() - today.getTime(); // Diferença em milissegundos
-    const differenceInDays = Math.ceil(differenceInTime / (1000 * 3600 * 24)); // Converte para dias
-    return differenceInDays;
-  };
 
   useEffect(() => {
     // Data de exemplo, pode ser alterada conforme a necessidade
